Extend StandardEject sequence to include LUN 1
this is for eject handling of dongles based on
Mediatek's cellular baseband processors which has a TF/SD card reader on
LUN 0 of the USB storage interface and the virtual cd-rom (which we want
to eject) on LUN 1. The previous StandardEject sequence did only work
for LUN 0, the new sequence will work for both LUN's.

Signed-off-by: Benjamin Berg <benjamin@sipsolutions.net>
Signed-off-by: René van Weert <r.vanweert@sowifi.com>
Reviewed-by: Lars Melin <larsm17@gmail.com>